+++ datacontract/engines/soda/check_soda_execute.py
@@ -30,7 +30,7 @@ def check_soda_execute(
     run.log_info("Running engine soda-core")
     scan = Scan()
 
-    if server.type == "s3" or server.type == "local":
+    if server.type in ["s3", "azure", "local"]:
         if server.format in ["json", "parquet", "csv"]:
             con = get_duckdb_connection(data_contract, server)
             scan.add_duckdb_connection(duckdb_connection=con, data_source_name=server.type)